LayoutTests/storage/indexeddb/basics-expected.txt

Test IndexedDB's basics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'indexedDB' in window is true
PASS indexedDB == null is false
indexedDB.open('name', 'description')
PASS 'onsuccess' in result is true
PASS 'onerror' in result is true
PASS 'readyState' in result is true
An event should fire shortly...

Success event fired:
PASS 'result' in event is true
PASS 'code' in event is false
PASS 'message' in event is false
PASS 'source' in event is true
PASS event.source != null is true
PASS 'onsuccess' in event.target is true
PASS 'onerror' in event.target is true
PASS 'readyState' in event.target is true
PASS event.target.readyState is event.target.DONE

PASS successfullyParsed is true

TEST COMPLETE
